+++ common/src/main/java/net/mehvahdjukaar/vista/integration/CompatHandler.java
@@ -27,6 +27,7 @@ public class CompatHandler {
     public static final boolean FLASHBACK = PlatHelper.isModLoaded("flashback");
     public static final boolean VEIL = PlatHelper.isModLoaded("veil");
     public static final boolean VAMPIRISM = PlatHelper.isModLoaded("vampirism");
+    public static final boolean SUPERNATURAL = PlatHelper.isModLoaded("supernatural");
     public static final boolean WATERMEDIA = isWatermediaCompatible();
 
     private static boolean isWatermediaCompatible() {